:recycle: #187 - refactor: create separate (sub)pages for editing and processing a review of a destruction list
